// frontend/src/store/thunks/interfaces/ICharacter.ts
export interface ICharacter {
  birth_year: string
  created: string
  edited: string
  eye_color: string
  films: string[]
  gender: string
  hair_color: string
  height: string
  homeworld: string
  mass: string
  name: string
  skin_color: string
  species: []
  starships: string[]
  url: string
  vehicles: string[]
}